Proof of concept to implement strict ordering of fieldIDs
Uses an alternative approach then what was previously discussed
regarding binary encoding of fieldIDs when serializing data store
keys.

Implements a "string based" strict ordering, instead of a "binary"
based approach. By applying a padding of zeros, and limiting the
stringified fieldID to 3 characters will gurantee ordering.

Basically, "001" < "002" < "011", where as before
"1" < "11" < "2".

This has the effect that we can only have 999 fields per document.

+++ db/fetcher/errors.go
@@ -25,6 +25,7 @@ const (
 	errVFetcherFailedToDecodeNode   string = "(version fetcher) failed to decode protobuf"
 	errVFetcherFailedToGetDagLink   string = "(version fetcher) failed to get node link from DAG"
 	errFailedToGetDagNode           string = "failed to get DAG Node"
+	// errSeekTargetBeforeCurrentState string = "cannot seek to target that is behind the current state"
 )
 
 var (
@@ -42,7 +43,7 @@ var (
 )
 
 // NewErrFieldIdNotFound returns an error indicating that the given FieldId was not found.
-func NewErrFieldIdNotFound(fieldId uint32) error {
+func NewErrFieldIdNotFound(fieldId uint16) error {
 	return errors.New(errFieldIdNotFound, errors.NewKV("FieldId", fieldId))
 }
